requirements.txt
absl-py==2.1.0
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
docopt==0.6.2
flatbuffers==24.3.25
fonttools==4.51.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
jax==0.4.26
jaxlib==0.4.26
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.8.4
mediapipe==0.10.13
ml-dtypes==0.4.0
moviepy==1.0.3
nltk==3.8.1
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pillow==10.3.0
proglog==0.1.10
protobuf==4.25.3
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytube==10.9.3
pytube3==9.6.4
pytubefix==5.4.2
regex==2024.4.28
requests==2.31.0
scipy==1.13.0
six==1.16.0
sounddevice==0.4.6
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
webvtt-py==0.4.6
youtube-dl==2021.12.17
youtube-transcript-api==0.6.2
